config(toml): move build config into manifest
