+++ src/lib.rs
@@ -23,7 +23,6 @@ extern crate toml;
 pub mod bindgen;
 pub mod build;
 pub mod command;
-pub mod config;
 pub mod emoji;
 pub mod error;
 pub mod logger;
